Extracting `usePersistedSetting` hook for reuse.
