+++ graylog2-web-interface/src/components/perspectives/contexts/PerspectivesProvider.tsx
@@ -16,15 +16,11 @@
  */
 import * as React from 'react';
 import type { PropsWithChildren } from 'react';
-import { useCallback, useContext, useMemo, useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 
 import usePluginEntities from 'hooks/usePluginEntities';
-import { useStore } from 'stores/connect';
-import { CurrentUserStore } from 'stores/users/CurrentUserStore';
-import UserPreferencesContext from 'contexts/UserPreferencesContext';
-import { PreferencesStore } from 'stores/users/PreferencesStore';
-import Store from 'logic/local-storage/Store';
 import type { Perspective } from 'components/perspectives/types';
+import usePersistedSetting from 'hooks/usePersistedSetting';
 
 import PerspectivesContext from './PerspectivesContext';
 
@@ -32,33 +28,6 @@ const DEFAULT_PERSPECTIVE = 'default';
 
 const findPerspective = (perspectives: Array<Perspective>, perspectiveId: string) =>
   perspectives.find(({ id }) => id === perspectiveId);
-
-const usePersistedSetting = (settingKey: string) => {
-  const { userIsReadOnly, username } = useStore(CurrentUserStore, (userStore) => ({
-    username: userStore.currentUser?.username,
-    userIsReadOnly: userStore.currentUser?.read_only ?? true,
-  }));
-
-  const userPreferences = useContext(UserPreferencesContext);
-  const setting = userIsReadOnly ? Store.get(settingKey) : userPreferences[settingKey];
-
-  const setSetting = useCallback(
-    (newSetting: string) => {
-      if (userIsReadOnly) {
-        Store.set(settingKey, newSetting);
-
-        return Promise.resolve();
-      }
-
-      const nextPreferences = { ...userPreferences, [settingKey]: newSetting };
-
-      return PreferencesStore.saveUserPreferences(username, nextPreferences);
-    },
-    [settingKey, userIsReadOnly, userPreferences, username],
-  );
-
-  return useMemo(() => [setting, setSetting], [setSetting, setting]);
-};
 
 const useActivePerspectiveState = (availablePerspectives: Array<Perspective>) => {
   const [persistedPerspective, setPersistedPerspective] = usePersistedSetting('perspective');
